pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexus"
version = "0.1.0"
description = "Local-first Personal AI OS"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115,<1.0",
    "pydantic>=2.10,<3.0",
    "aiosqlite>=0.20,<1.0",
    "numpy>=2.0,<3.0",
    "httpx>=0.28,<1.0",
    "pyyaml>=6.0,<7.0",
    "ollama>=0.4,<1.0",
    "groq>=0.13,<2.0",
    "anthropic>=0.40,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<10.0",
    "pytest-asyncio>=0.24,<1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["nexus"]
